[WOOR-224] chore: 이미지 모듈 분리
