--- src/main/java/com/musseukpeople/woorimap/common/config/AwsConfig.java
@@ -1,50 +1,0 @@
-package com.musseukpeople.woorimap.common.config;
-
-import javax.annotation.PostConstruct;
-
-import org.springframework.beans.factory.annotation.Value;
-import org.springframework.context.annotation.Bean;
-import org.springframework.context.annotation.Configuration;
-import org.springframework.context.annotation.Profile;
-
-import com.amazonaws.auth.AWSCredentials;
-import com.amazonaws.auth.AWSCredentialsProvider;
-import com.amazonaws.auth.AWSStaticCredentialsProvider;
-import com.amazonaws.auth.BasicAWSCredentials;
-import com.amazonaws.services.s3.AmazonS3Client;
-import com.amazonaws.services.s3.AmazonS3ClientBuilder;
-
-@Profile("!test")
-@Configuration
-public class AwsConfig {
-
-    @Value("${cloud.aws.credentials.access-key}")
-    private String accessKey;
-
-    @Value("${cloud.aws.credentials.secret-key}")
-    private String secretKey;
-
-    @Value("${cloud.aws.region.static}")
-    private String region;
-
-    private AWSCredentials awsCredentials;
-
-    @PostConstruct
-    public void init() {
-        awsCredentials = new BasicAWSCredentials(accessKey, secretKey);
-    }
-
-    @Bean
-    public AWSCredentialsProvider awsCredentialsProvider() {
-        return new AWSStaticCredentialsProvider(awsCredentials);
-    }
-
-    @Bean
-    public AmazonS3Client amazonS3Client() {
-        return (AmazonS3Client)AmazonS3ClientBuilder.standard()
-            .withRegion(region)
-            .withCredentials(new AWSStaticCredentialsProvider(awsCredentials))
-            .build();
-    }
-
-}
