+++ woorimap-image/src/main/java/com/musseukpeople/woorimapimage/WoorimapImageApplication.java
@@ -1,0 +1,13 @@
+package com.musseukpeople.woorimapimage;
+
+import org.springframework.boot.SpringApplication;
+import org.springframework.boot.autoconfigure.SpringBootApplication;
+
+@SpringBootApplication
+public class WoorimapImageApplication {
+
+    public static void main(String[] args) {
+        SpringApplication.run(WoorimapImageApplication.class, args);
+    }
+
+}
