+++ woorimap-image/src/main/java/com/musseukpeople/woorimapimage/common/model/ApiResponse.java
@@ -1,0 +1,16 @@
+package com.musseukpeople.woorimapimage.common.model;
+
+import lombok.AccessLevel;
+import lombok.Getter;
+import lombok.NoArgsConstructor;
+
+@Getter
+@NoArgsConstructor(access = AccessLevel.PRIVATE)
+public class ApiResponse<T> {
+
+    private T data;
+
+    public ApiResponse(T data) {
+        this.data = data;
+    }
+}
