+++ woorimap-image/src/main/java/com/musseukpeople/woorimapimage/common/util/FileUtil.java
@@ -1,0 +1,14 @@
+package com.musseukpeople.woorimapimage.common.util;
+
+import java.io.File;
+
+import lombok.AccessLevel;
+import lombok.NoArgsConstructor;
+
+@NoArgsConstructor(access = AccessLevel.PRIVATE)
+public class FileUtil {
+
+    public static boolean removeFile(File file) {
+        return file.delete();
+    }
+}
