+++ woorimap-image/src/main/java/com/musseukpeople/woorimapimage/image/application/ImageService.java
@@ -1,0 +1,12 @@
+package com.musseukpeople.woorimapimage.image.application;
+
+import java.util.List;
+
+import org.springframework.web.multipart.MultipartFile;
+
+public interface ImageService {
+
+    List<String> uploadImages(Long id, List<MultipartFile> files);
+
+    String uploadImage(Long id, MultipartFile file);
+}
