+++ woorimap-image/src/main/java/com/musseukpeople/woorimapimage/image/exception/ConvertFileFailedException.java
@@ -1,0 +1,13 @@
+package com.musseukpeople.woorimapimage.image.exception;
+
+import java.text.MessageFormat;
+
+import com.musseukpeople.woorimapimage.common.exception.ErrorCode;
+
+public class ConvertFileFailedException extends FileException {
+    private static final String ERROR_MESSAGE_FORMAT = "MultiPartFile 변환에 실패했습니다. 파일명 : {0}";
+
+    public ConvertFileFailedException(String fileName, ErrorCode errorCode) {
+        super(MessageFormat.format(ERROR_MESSAGE_FORMAT, fileName), errorCode);
+    }
+}
