+++ woorimap-image/src/main/java/com/musseukpeople/woorimapimage/image/presentation/auth/login/Login.java
@@ -1,0 +1,14 @@
+package com.musseukpeople.woorimapimage.image.presentation.auth.login;
+
+import java.lang.annotation.ElementType;
+import java.lang.annotation.Retention;
+import java.lang.annotation.RetentionPolicy;
+import java.lang.annotation.Target;
+
+import io.swagger.v3.oas.annotations.Parameter;
+
+@Target(ElementType.PARAMETER)
+@Retention(RetentionPolicy.RUNTIME)
+@Parameter(hidden = true)
+public @interface Login {
+}
