+++ woorimap-image/src/main/java/com/musseukpeople/woorimapimage/image/presentation/auth/login/LoginMember.java
@@ -1,0 +1,15 @@
+package com.musseukpeople.woorimapimage.image.presentation.auth.login;
+
+import lombok.Getter;
+
+@Getter
+public class LoginMember {
+
+    private final Long id;
+    private final String accessToken;
+
+    public LoginMember(Long id, String accessToken) {
+        this.id = id;
+        this.accessToken = accessToken;
+    }
+}
