+++ woorimap-image/src/main/resources/application-test.yml
@@ -1,0 +1,22 @@
+spring:
+  servlet:
+    multipart:
+      max-file-size: 10MB
+      max-request-size: 10MB
+
+cloud:
+  aws:
+    credentials:
+      access-key: test
+      secret-key: test
+    s3:
+      bucket: test
+    region:
+      static: ap-northeast-2
+    stack:
+      auto: false
+
+woorimap:
+  api:
+    host: http://test
+    port: :8080
