+++ woorimap-image/src/main/resources/application.yml
@@ -1,0 +1,49 @@
+server:
+  port: 8081
+
+spring:
+  profiles:
+    active: local
+
+  servlet:
+    multipart:
+      max-file-size: 10MB
+      max-request-size: 10MB
+
+cloud:
+  aws:
+    credentials:
+      access-key: ${AWS_ACCESS_KEY_ID}
+      secret-key: ${AWS_SECRET_ACCESS_KEY}
+    s3:
+      bucket: wooriemap
+    region:
+      static: ap-northeast-2
+    stack:
+      auto: false
+
+woorimap:
+  api:
+    host: http://localhost
+    port: :8080
+
+---
+spring:
+  config:
+    activate:
+      on-profile: dev
+
+woorimap:
+  api:
+    host: https://dev.woorimap.p-e.kr/
+    port:
+---
+spring:
+  config:
+    activate:
+      on-profile: prod
+
+woorimap:
+  api:
+    host: https://prod.woorimap.p-e.kr/
+    port:
